pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "gce-reference-prototype"
version = "0.1.0"
description = "Reference prototype and reproducibility artifact for an ML-KEM-768-based incompleteness-secured commitment framework"
readme = "README.md"
requires-python = ">=3.10"
license = {file = "LICENSE"}
authors = [
  {name = "Kevin Henry Miller", email = "kevin@qbondnetwork.com"}
]
keywords = ["cryptography", "post-quantum", "ml-kem", "commitment", "research-artifact"]
classifiers = [
  "Development Status :: 3 - Alpha",
  "Intended Audience :: Science/Research",
  "Programming Language :: Python :: 3",
  "Topic :: Security :: Cryptography"
]
dependencies = [
  "keeper-mlkem==1.0.2",
  "fips203==0.4.3"
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
